requirements.txt
connexion >= 2.14.2
connexion[swagger-ui] >= 2.14.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
mysql-connector-python==8.0.29
pymysql==1.0.2
sqlalchemy==1.4.47
Flask-SQLAlchemy==2.5.1
Flask==2.2.3
Werkzeug==2.2.2
flask-marshmallow==0.14.0